Добавлен сервис для депозитов

+++ basic_starter/src/main/java/org/example/starter/dto/RequestDto.java
@@ -32,5 +32,5 @@ public class RequestDto {
     /**
      * Идентификатор депозита
      */
-    private Integer depositId;
+    private Long depositId;
 }
