+++ customerservice/src/main/java/org/example/customerservice/entity/Request.java
@@ -7,6 +7,9 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import java.time.LocalDate;
 
+/**
+ * Объект данных запроса на открытие депозита
+ */
 @Entity
 @Table(name = "requests", schema = "customers")
 @Data
@@ -14,18 +17,30 @@ import java.time.LocalDate;
 @NoArgsConstructor
 public class Request {
 
+    /**
+     * Идентификатор запроса
+     */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_request")
     private Long id;
 
+    /**
+     * Идентификатор клиента
+     */
     @ManyToOne
     @JoinColumn(name = "customer_id", nullable = false)
     private Customer customer;
 
+    /**
+     * Дата отправки запроса
+     */
     @Column(name = "request_date", nullable = false)
     private LocalDate date;
 
+    /**
+     * Идентификатор депозита
+     */
     @Column(name = "deposits_id")
-    private Integer depositId;
+    private Long depositId;
 }
